requirements.txt
addict==2.4.0
aistudio-sdk==0.3.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astor==0.8.1
attrdict==2.0.1
babel==2.18.0
bce-python-sdk==0.9.68
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.0
crc32c==2.8
crcmod==1.7
cryptography==46.0.6
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
decorator==5.2.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.14.0
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
fonttools==4.60.2
fsspec==2025.10.0
future==1.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
ImageIO==2.37.2
imagesize==1.5.0
imgaug==0.4.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==0.10.0
kiwisolver==1.4.7
lazy-loader==0.5
lmdb==2.2.0
lxml==6.0.2
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mmcv==2.1.0
-e git+https://github.com/open-mmlab/mmdetection.git@cfd5d3a985b0249de009b67d04f37263e11cdf3d#egg=mmdet
mmengine==0.10.7
model-index==0.1.11
modelscope==1.35.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
ninja==1.13.0
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.5
openxlab==0.1.3
opt-einsum==3.3.0
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
paddleocr==2.7.0.3
paddlepaddle==2.6.2
paddlex==3.4.3
pandas==2.3.3
pdf2docx==0.5.8
pillow==11.3.0
platformdirs==4.4.0
premailer==3.10.0
prettytable==3.16.0
protobuf==3.20.2
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycocotools==2.0.11
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
PyMuPDF==1.20.2
pyparsing==3.3.2
pypdfium2==5.6.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2023.4
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
regex==2026.1.15
reportlab==4.0.9
requests==2.28.2
rich==13.4.2
ruamel.yaml==0.19.1
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
terminaltables==3.1.10
tifffile==2024.8.30
tomli==2.4.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.2
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==1.26.20
uvicorn==0.39.0
visualdl==2.5.3
wcwidth==0.6.0
Werkzeug==3.1.8
yapf==0.43.0
zipp==3.23.0
